\name{packagelabthree-package}
\alias{packagelabthree-package}
\alias{packagelabthree}
\docType{package}
\title{
\packageTitle{packagelabthree}
}
\description{
\packageDescription{packagelabthree}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{packagelabthree}
\packageIndices{packagelabthree}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{packagelabthree}

Maintainer: \packageMaintainer{packagelabthree}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
