% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packagelabthree.R
\docType{package}
\name{packagelabthree}
\alias{packagelabthree}
\alias{packagelabthree-package}
\title{packagelabthree: A package containing the euclidian algorithm that finds the greatest common divisor of two integers.}
\description{
packagelabthree: A package containing the euclidian algorithm that finds the greatest common divisor of two integers.
}

