% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/euclidean.R, R/packagelabthree.R
\docType{package}
\name{euclidean}
\alias{euclidean}
\alias{euclidean-package}
\title{Finds the greatest common divisor of two numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar or integer.}

\item{b}{A numeric scalar or integer.}
}
\value{
A numeric value which is the greatest common divisor of a and b.
}
\description{
The euclidian algorithm finds the greatest common divisor of two numbers.
More about the algorithm can be found at https://en.wikipedia.org/wiki/Euclidean_algorithm.

euclidean: A package using the euclidian algorithm to find the greatest common divisor of two integers.
}

